# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Salsyx API",
    "description": "Search and preservation platform for public GitHub repositories. Nothing open-source should disappear forever.",
    "version": "0.1.0",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "/api/v1", "description": "Default base path" }
  ],
...
  "paths": {
...
    "/repo/{owner}/{repo}": {
      "get": {
        "tags": ["repository"],
        "summary": "Resolve a repository",
        "description": "Checks GitHub first; if deleted, falls back to the Salsyx archive database.",
        "operationId": "resolveRepository",
        "parameters": [
          { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Resolution result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RepoResponse" }
              }
            }
          },
          "404": {
            "description": "Not found on GitHub and not archived",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorBody" }
              }
            }
          }
        }
      }
    },
...
    "/download/{id}": {
      "get": {
        "tags": ["archive"],
        "summary": "Download archived blob",
        "description": "Streams the archived snapshot with integrity verification.",
        "operationId": "downloadArchive",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Binary archive stream"
          }
        }
      }
    },
...
    "/archive": {
      "post": {
        "tags": ["archive"],
        "summary": "Enqueue an archive job",
        "operationId": "createArchive",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateArchiveRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Archive job queued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateArchiveResponse" }
              }
            }
          }
        }
      }
    },
    "/refresh": {
      "post": {
        "tags": ["repository"],
        "summary": "Force-refresh a repository",
        "operationId": "refreshRepository",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RefreshResponse" }
              }
            }
          }
        }
      }
    },
    "/repo/{owner}/{repo}/readme": {
      "get": {
        "tags": ["repository"],
        "summary": "Get the README",
        "description": "Returns the default-branch README, preferring the preserved snapshot and falling back to a live GitHub fetch.",
        "operationId": "getReadme",
        "parameters": [
          { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "README content",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReadmeResponse" }
              }
            }
          },
          "404": {
            "description": "No README available",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorBody" }
              }
            }
          }
        }
      }
    },
    "/repo/{owner}/{repo}/archives": {
      "get": {
        "tags": ["repository"],
        "summary": "Archive history for a repository",
        "operationId": "repoArchiveHistory",
        "parameters": [
          { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Archives, newest first",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HistoryResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RepoResponse": {
        "type": "object",
        "properties": {
          "source": { "type": "string", "enum": ["github", "salsyx"] },
          "status": { "type": "string", "enum": ["live", "archived", "not_found"] },
          "repository": { "type": "object" },
          "archive": { "type": "object" },
          "download_url": { "type": "string" },
          "message": { "type": "string" }
        }
      },
...
      "CreateArchiveRequest": {
        "type": "object",
        "required": ["full_name"],
        "properties": {
          "full_name": { "type": "string", "description": "owner/repo" }
        }
      },
      "CreateArchiveResponse": {
        "type": "object",
        "properties": {
          "archive_id": { "type": "string", "format": "uuid" },
          "status": { "type": "string" },
          "message": { "type": "string" }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["full_name"],
        "properties": {
          "full_name": { "type": "string", "description": "owner/repo" }
        }
      },
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "full_name": { "type": "string" },
          "status": { "type": "string" },
          "source": { "type": "string" },
          "archive_id": { "type": "string", "format": "uuid" }
        }
      },
      "ReadmeResponse": {
        "type": "object",
        "properties": {
          "source": { "type": "string", "enum": ["salsyx", "github"] },
          "readme": { "type": "string" },
          "html_url": { "type": "string" }
        }
      },
...
      "HistoryArchive": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "commit_ref": { "type": "string" },
          "commit_count": { "type": "integer" },
          "checksum": { "type": "string" },
          "size_bytes": { "type": "integer" },
          "status": { "type": "string" },
          "archived_at": { "type": "string", "format": "date-time" }
        }
      },
      "HistoryResponse": {
        "type": "object",
        "properties": {
          "owner": { "type": "string" },
          "repo": { "type": "string" },
          "archives": { "type": "array", "items": { "$ref": "#/components/schemas/HistoryArchive" } }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "detail": { "type": "string" }
        }
      }
    }
  }
}